+++ src/shared/types/index.ts
@@ -4,5 +4,6 @@ export * from "./keyboard";
 export * from "./language";
 export * from "./number";
 export * from "./option";
+export * from "./position";
 export * from "./sort";
 export * from "./toolbar";